Initial implementation of G-code validator, which is intended to validate G-code when its uploaded to the designer portal.
However, it is important to note that we do NOT want to make it difficult to upload G-code.  We don't want designers getting frustrated and walking away because they can't get their G-code uploaded.  The way this should really work is that the tool should warn the designer that there's a problem with their G-code, but automatically fix it for them.