Major project update.  This includes:
- Wholesale cleanup/removal of resources that are no longer being used
- Implementation of Toggle's for the iOS UI
- Creation of FBX files for iOS UI and Panerai
- Addition of Panerai to the project
  - As an aside, the iOS UI should also eventually come from an FBX.  This is helpful because it prevents unnecessary processing of models in the Blender source files during importation.  Also, the FBX files may be treated as prefabs which will maintain their association with the data in the FBX file (e.g. if you change the name of an element in the Blender source, it will propagate through the FBX file into your scene).  This also prevents Unity from spamming your project with a bunch of unused materials, thereby giving us a reason to actually allow Unity to create materials upon import.
- Implementation of optical light tracking with gyroscopic motion augmentation
  - This implementation is insufficient as we are currently only tracking the scene from the front camera.  There are points where, as you rotate the device, you may observe the sun light stop moving relative to the device.  Ultimately, we need to track both front and rear cams (currently not possible with iOS), or as a degraded solution, implement some sort of mechanism for allowing the sun to move beyond the bounds of the front image.

This work included a ton of time spent trying to debug specular highlights on iOS.  There is no configuration variation (e.g. changing from Metal to OpenGL ES 3, etc) that will cause the specular highlight striations to go away.  I also modified the standard shader to figure out whether there was something stupid like arithmetic overflows happening, but I don't believe that's the case.  Rather, I think what's happening is the use of half's in the shader is taking effect while in use on iOS (half's are treated as full float's on desktop), which is causing the resolution of some of the directional data to be reduced exponentially.  The next thing for me to do here would be to go ahead and force use of full floats on iOS.  However, I've deferred that as a bridge too far.