- Change the PolygonSet "normal" that previously represented the normal of the plane upon which a 2D arrangement of points resided (and was used exactly nowhere), to instead use orientation (a Quaternion). - Add support to the Set Rotation node to set this orientation. - Change the Mesh Display node to apply the rotation when generating the displayable for of polygonal meshes.
As an aside, it occurred to me while doing this that the BlitItMesh class operates on triangle meshes only, yet the polygonal stuff is in the Shapes namespace. It certainly isn't functionally necessary to unify these, but it would seem that some consistency between these might be in order. To be clear, I am not at all saying that BlitItMesh should support polygonal meshes.  That remains to be seen.  In my opinion, the ability for Unity to render line meshes using the Mesh class is a bit discordant.  I think it probably makes more sense to keep polygonal data separate from triangle data, as operating on this data is inherently different (3D vs 2D).  With sufficient ability to translate between them when necessary (i.e. trangulating polygonal data, or even converting a triangle mesh into polygonal data -- a concept that seems strange to me as I write this -- the some similar naming would be warranted.