Correct Unity's absolutely ridiculous behavior of not saving changes to properties defined as "default overrides."
They're basically using "prefab" like a "class" instead of as a prefab.  That is, people are using a prefab as a base template of objects in a scene, expecting that when they modify the prefab, all of the instances of that prefab in the scene are updated dynamically.  This makes almost zero sense in the case that you modify the prefab transform (as it would collapse all instances of that prefab onto the same position, for example), so their solution was to give certain properties (the transform) special treatment by preventing them from being saved when ApplyPrefabInstance() is called.  All of this is absolutely, 100% ridiculous, because it presupposes that a prefabrication of an object isn't a thing from which a copy is constructed (the copy then taking of a life of its own, disconnected from the parent), but has a dynamic influence of any instances derived from it.  Prefabs are prefabs.  Classes are classes.  Instances are instances.  These things are separate and distinct, and have no relation to one another.