Since the purpose of the mesh preview node is to render the mesh to the user, make sure it only happens once per frame.
This has brought to the forefront the question of how the UI of the node application should update vs how data should be processed.  After thinking about this for a while, I believe the way this should work is that UI updates must happen in near-constant time with only small amounts of data, therefore the main thread is essentially never blocked.  All data processing (any data which scales with size...e.g. O(n)) should be processed in background threads.  An implication of this is that it's possible that SetValue() should only ever be called in a background thread, as the number of SetValue() calls may grow with the size of the data being processed.